Intel(R) SGX DCAP 1.6 Release
Added APIs to configure file directory for DCAP quoting Enclave, quote
  provider library and quote verification library.
Fixed bugs.

+++ QuoteVerification/QVL/Src/AttestationLibrary/src/QuoteVerification/ByteOperands.cpp
@@ -52,8 +52,8 @@ uint32_t swapBytes(uint32_t val)
 uint16_t toUint16(uint8_t leftMostByte, uint8_t rightMostByte)
 {
     uint32_t ret = 0;
-    ret |= static_cast<uint16_t>(rightMostByte);
-    ret |= (static_cast<uint16_t>(leftMostByte) << 8) & 0xff00;
+    ret |= static_cast<uint32_t>(rightMostByte);
+    ret |= (static_cast<uint32_t>(leftMostByte) << 8) & 0xff00;
     return static_cast<uint16_t>(ret);
 }
